refactor: rename table_driven_rule_test to table_driven_rule_test_base
